app starting, created new main page, implemented new errors, spring h2 db connection, mapping to entities, commented old code for rework, edited entity properties

+++ src/main/java/es/ulpgc/bowling/entity/BaseEntity.java
@@ -9,13 +9,13 @@ public class BaseEntity implements Serializable {
     @Id
     @Column
     @GeneratedValue(strategy=GenerationType.AUTO)
-    Long id;
+    int id;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
